requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
astor=0.8.1=pypi_0
babel=2.9.1=pypi_0
backports-entry-points-selectable=1.1.1=pypi_0
bce-python-sdk=0.8.64=pypi_0
cachetools=4.2.4=pypi_0
certifi=2021.10.8=pypi_0
cfgv=3.0.0=pypi_0
charset-normalizer=2.0.8=pypi_0
click=7.1.2=pypi_0
colorama=0.4.4=pypi_0
cssselect=1.1.0=pypi_0
cssutils=2.3.0=pypi_0
cython=0.29.24=pypi_0
dataclasses=0.8=pypi_0
distlib=0.3.3=pypi_0
easydict=1.9=pypi_0
et-xmlfile=1.1.0=pypi_0
fasttext=0.9.1=pypi_0
filelock=3.4.0=pypi_0
filterpy=1.4.5=pypi_0
flake8=4.0.1=pypi_0
flask=2.0.2=pypi_0
flask-babel=2.0.0=pypi_0
flatbuffers=2.0=pypi_0
future=0.18.2=pypi_0
identify=1.6.2=pypi_0
idna=3.3=pypi_0
imgaug=0.4.0=pypi_0
importlib-metadata=4.2.0=pypi_0
importlib-resources=5.4.0=pypi_0
intel-openmp=2022.1.0=h57928b3_3787
itsdangerous=2.0.1=pypi_0
jinja2=3.0.3=pypi_0
joblib=1.1.0=pypi_0
lap=0.4.0=py36h5076640_1002
libblas=3.9.0=8_mkl
libcblas=3.9.0=8_mkl
liblapack=3.9.0=8_mkl
lmdb=1.2.1=pypi_0
lxml=4.6.4=pypi_0
markupsafe=2.0.1=pypi_0
mccabe=0.6.1=pypi_0
mkl=2020.4=hb70f87d_311
motmetrics=1.2.5=pypi_0
networkx=2.5.1=pypi_0
nodeenv=1.6.0=pypi_0
numpy=1.19.3=pypi_0
onnx=1.6.0=pypi_0
onnx-simplifier=0.3.6=pypi_0
onnxoptimizer=0.2.6=pypi_0
onnxruntime=1.9.0=pypi_0
opencv-contrib-python=4.4.0.46=pypi_0
opencv-python=4.5.4.60=pypi_0
openpyxl=3.0.9=pypi_0
packaging=21.3=pypi_0
paddle2onnx=0.8.2=pypi_0
paddleclas=0.0.0=pypi_0
paddleocr=2.3.0.2=pypi_0
paddlepaddle-gpu=2.2.0=pypi_0
pandas=0.25.3=pypi_0
pillow=8.4.0=pypi_0
pip=21.3.1=pyhd8ed1ab_0
platformdirs=2.4.0=pypi_0
pre-commit=2.1.1=pypi_0
premailer=3.10.0=pypi_0
protobuf=3.19.1=pypi_0
pybind11=2.8.1=pypi_0
pyclipper=1.3.0=pypi_0
pycodestyle=2.8.0=pypi_0
pycryptodome=3.12.0=pypi_0
pyflakes=2.4.0=pypi_0
pyparsing=2.4.7=pypi_0
pyqt5=5.15.4=pypi_0
pyqt5-plugins=5.15.4.2.2=pypi_0
pyqt5-qt5=5.15.2=pypi_0
pyqt5-sip=12.9.1=pypi_0
pyqt5-tools=5.15.4.3.2=pypi_0
python=3.6.0=0
python-dateutil=2.8.2=pypi_0
python-dotenv=0.20.0=pypi_0
python-levenshtein=0.12.2=pypi_0
python_abi=3.6=2_cp36m
pytz=2021.3=pypi_0
pyyaml=6.0=pypi_0
qt5-applications=5.15.2.2.2=pypi_0
qt5-tools=5.15.2.1.2=pypi_0
requests=2.26.0=pypi_0
scikit-learn=0.24.2=pypi_0
scipy=1.5.4=pypi_0
setuptools=36.4.0=py36_1
shapely=1.8.0=pypi_0
shellcheck-py=0.8.0.1=pypi_0
sip=6.5.1=pypi_0
six=1.16.0=pypi_0
threadpoolctl=3.1.0=pypi_0
toml=0.10.2=pypi_0
tornado=6.1=pypi_0
tqdm=4.62.3=pypi_0
typing-extensions=4.0.0=pypi_0
ucrt=10.0.20348.0=h57928b3_0
urllib3=1.26.7=pypi_0
vc=14=0
virtualenv=20.10.0=pypi_0
visualdl=2.2.2=pypi_0
vs2015_runtime=14.29.30037=h902a5da_5
werkzeug=2.0.2=pypi_0
wheel=0.37.0=pyhd8ed1ab_1
wincertstore=0.2=py36ha15d459_1006
xmltodict=0.13.0=pypi_0
zipp=3.6.0=pypi_0
